Fix compiler warnings in audio/NLP samples
Resolve all default-level GCC/Clang format-string warnings:
- audio_processing/gpt-sovits: cast std::chrono ms count to long long
  to match the %lld specifier (twice).
- natural_language_processing/g2p_en: same chrono cast for the
  benchmark print, and switch %d to %zu for std::vector::size().
- natural_language_processing/sentence_transformers,
  natural_language_processing/multilingual-e5: switch the second
  format spec from %d to %zu for std::vector::size().

After these changes, both `cmake -DWITH_OPENCV=ON .` and
`cmake -DWITH_OPENCV=OFF .` builds produce zero warnings on GCC.

+++ audio_processing/gpt-sovits/gpt-sovits.cpp
@@ -444,7 +444,7 @@ static AILIATensor t2s_forward(AILIATensor ref_seq, AILIATensor text_seq, AILIAT
 		}
 
 		if (benchmark){
-            PRINT_OUT("ailia processing time %lld ms\n",  std::chrono::duration_cast<std::chrono::milliseconds>(end2 - start2).count());
+            PRINT_OUT("ailia processing time %lld ms\n",  (long long)std::chrono::duration_cast<std::chrono::milliseconds>(end2 - start2).count());
 		}
 
 		if (stop){
@@ -636,7 +636,7 @@ int main(int argc, char **argv)
 	status = recognize_from_audio(ailia);
 	auto end2 = std::chrono::high_resolution_clock::now();
 	if (benchmark){
-		PRINT_OUT("total processing time %lld ms\n",  std::chrono::duration_cast<std::chrono::milliseconds>(end2 - start2).count());
+		PRINT_OUT("total processing time %lld ms\n",  (long long)std::chrono::duration_cast<std::chrono::milliseconds>(end2 - start2).count());
 	}
 
 	for (int i = 0; i < MODEL_N; i++){
